requirements.txt
aescipher==6.0.0
aniso8601==9.0.1
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
cryptography==43.0.3
distlib==0.3.9
dj-database-url==0.5.0
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.9.0
docutils==0.21.2
easyrsa==3.1.0
encryptedsocket==3.0.0
filelock==3.16.1
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.8
importlib_resources==6.4.4
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.5.0
ldjango==8.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
munch==4.0.0
nh3==0.2.19
numpy==2.1.3
omnitools==0.0.268
packaging==24.1
pillow==11.0.0
pipenv==2024.4.0
pkginfo==1.12.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.22
pycryptodome==3.21.0
pyftpdlib==2.0.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythoncgi==0.17.0
pytz==2024.1
readme_renderer==44.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.20.0
setuptools==75.3.0
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.2
threadwrapper==0.14.0
tqdm==4.67.1
twine==6.0.1
typing_extensions==4.12.2
tzlocal==5.2
unencryptedsocket==2.2.0
urllib3==2.2.2
virtualenv==20.27.1
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.45.1
whitenoise==6.8.2
